# ub-device-manager/ub_device_manager/openapi.yaml
...
openapi: 3.0.3
info:
  title: UB Device Manager
  version: 1.0.0
  description: 仅在本地节点运行的 UDS HTTP 服务
servers:
  - url: http://localhost
...
paths:
  /ssu-ns/{name}:
    get:
      summary: 获取存储空间命名信息
      description: 查询指定存储空间下各命名空间的容量使用情况, 包括总容量和已用容量
      operationId: show_ssu_ns_status
      tags:
        - SSU
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: 存储空间标识


      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: array
                description: 命名空间统计信息列表
                items:
                  $ref: '#/components/schemas/SsuNsStatus'
  /ssu-ns/connect-info:
    get:
      summary: 查询指定存储空间在指定 VFE 上的 NVMe 连接信息
      description: 查询指定存储空间在指定 VFE 上的 NVMe 连接信息, 包括子系统 NQN、Host NQN、命名空间 ID 等
      operationId: get_ns_connnect_info
      tags:
        - SSU
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: 存储空间标识
        - name: vfe_guid
          in: query
          required: false
          schema:
            type: string
          description: vfe_guid
          example: "0000-0-00000000"
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: array
                description: 命名空间统计信息列表
                items:
                  $ref: '#/components/schemas/NsConnectInfo'
...
  /ssu/{name}:
    get:
      summary: 根据名称获取已分配的存储空间信息
      description: 根据名称获取已分配的存储空间信息, 包括命名空间列表、容量、LBA 格式和使用类型等
      operationId: show_ssu_alloc_info
      tags:
        - SSU
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: SSU 的名称
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsuInfo'
    delete:
      summary: 释放 SSU 空间
      description: 根据给定的 SSU 名称，释放其对应的空间资源
      operationId: free_ssu
      tags:
        - SSU
      parameters:
        - name: name
          in: path
          required: true
          description: SSU 的名称
          schema:
            type: string
            example: "ssu-volume-001"
      responses:
        '200':
          description: 释放SSU成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /ssu/alloc:
    post:
      summary: 分配 SSU 存储空间
      description: 根据请求参数分配指定数量和大小的命名空间, 支持顺序分配和分布式分配两种策略
      operationId: ssu_alloc
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsuAllocSpaceReq'
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsuInfo'
  /ssu/add-perm:
    post:
      summary: 添加 SSU 访问权限
      description: 为指定的 SSU 卷添加特定主机的 NQN 访问权限
      operationId: add_ssu_perm
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsuPermReq'
      responses:
        '200':
          description: 权限添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: 请求参数错误（如 NQN 格式不正确、SSU 名称为空）
        '404':
          description: 未找到指定的 SSU 资源
  /ssu/rm-perm:
    post:
      summary: 删除 SSU 访问权限
      description: 为指定的 SSU 卷删除指定主机 NQN 访问权限
      operationId: remove_ssu_perm
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsuPermReq'
      responses:
        '200':
          description: 权限删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: 请求参数错误（如 NQN 格式不正确、SSU 名称为空）
        '404':
          description: 未找到指定的 SSU 资源
  /ssu/attach:
    post:
      summary: 挂载已分配的 SSU 存储空间
      description: 将指定的存储空间挂载到系统, 返回挂载后的设备路径
      operationId: ssu_attach
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsuSpaceReq'
      responses:
        '200':
          description: 挂载成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsuAttachResult'
  /ssu/detach:
    post:
      summary: 卸载已分配的 SSU 存储空间
      description: 将指定的存储空间从系统卸载, 释放设备占用
      operationId: ssu_detach
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsuSpaceReq'
      responses:
        '200':
          description: 卸载成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
  /ssu-vfe/bind:
    post:
      summary: 将指定的 VFE 绑定到总线
      description: 将指定的 VFE 绑定到总线，以实现虚机通过此 VFE 与物理机通信，不填 bus_guid 时默认分配
      operationId: bind_ssu_vfe_bus
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindSsuVfeReq'

      responses:
        '200':
          description: 权限添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /ssu-vfe/unbind:
    post:
      summary: 将指定的 VFE 与 bus_guid 解绑
      description: 传入 VFE 的 GUID，并封装为 VFE 对象以及通过查询 UBSE 找到 VFE 对应的 bus_guid ，组合后调用解绑接口，由于虚机仅能使用一个 ssu-vfe，因此参数为唯一 vfe_guid
      operationId: unbind_ssu_vfe
      tags:
        - SSU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnbindSsuVfeReq'
      responses:
        '200':
          description: 解绑 bus 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  schemas:
...
    NsConnectInfo:
      type: object
      description: Storage connection information.
      properties:
        src_eid:
          type: string
          default: ""
          description: Source EID.
        tgt_eid:
          type: string
          default: ""
          description: Target EID.
        tgt_nqn:
          type: string
          default: ""
          description: Target NQN address.
        host_nqn:
          type: string
          default: ""
          description: Host NQN.
        ns_uuid:
          type: string
          default: ""
          description: Namespace UUID.
        ns_id:
          type: integer
          default: 0
          description: Namespace ID.
    SsuNsStatus:
      type: object
      description: Namespace usage statistics.
      properties:
        ns_uuid:
          type: string
          default: ""
          description: Namespace UUID.
        ns_id:
          type: integer
          default: 0
          description: Namespace ID.
        total_size:
          type: integer
          default: 0
          description: Total namespace size.
        used_size:
          type: integer
          default: 0
          description: Used namespace size.
    BindSsuVfeReq:
      type: object
      description: Request to bind a VFE.
      required:
        - upi
        - vfe_guid
        - bus_guid
      properties:
        upi:
          type: integer
          description: Tenant isolation identifier.
        vfe_guid:
          type: string
          description: SSU VFE GUID.
          example: "0000-0-0000-00000000"
        bus_guid:
          type: string
          maxLength: 32
          description: >-
            Bus instance GUID. Can be empty (empty string means ubse creates a
            new VM bus instance internally); when non-empty, its length must be 32.
          example: ""
    UnbindSsuVfeReq:
      type: object
      description: Request to unbind a VFE.
      required:
        - upi
        - vfe_guid
      properties:
        upi:
          type: integer
          description: Tenant isolation identifier.
        vfe_guid:
          type: string
          description: SSU VFE GUID.
          example: "0000-0-0000-00000000"
...
    SsuPermReq:
      type: object
      required:
        - name
        - nqn
      properties:
        name:
          type: string
          description: SSU name.
          example: "ssu-volume-001"
        nqn:
          type: string
          description: NVMe host NQN identifier.
          example: "nqn.2014-08.org.nvmexpress:uuid:fedcba98-7654-3210-fedc-ba9876543210"
    SsuLbaFormat:
      type: integer
      description: >
        LBA format definition.
        FORMAT_512 = 512  # 512B
        FORMAT_4K = 4096  # 4K
      enum:
        - 512
        - 4096
    SsuAllocStrategy:
      type: integer
      description: >
        Allocation strategy:
        0: STRIPED (distributed strategy).
        1: LINEAR (sequential strategy).
        2: NORMAL.
      enum:
        - 0
        - 1
        - 2
    SsuAllocSpaceReq:
      type: object
      description: Storage allocation request.
      required:
        - name
        - ns_size_gb
        - ns_num
        - lba_format
        - strategy
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 48
          description: Request identifier, with a maximum length of 48 characters.
        ns_size_gb:
          type: integer
          format: int64
          minimum: 1
          description: >-
            Requested total capacity in GB. Must be greater than or equal to 1.
        ns_num:
          type: integer
          format: int32
          minimum: 1
          description: >-
            Number of namespaces. When ns_num is 1, strategy cannot be STRIPED (0).
        lba_format:
          $ref: '#/components/schemas/SsuLbaFormat'
        strategy:
          $ref: '#/components/schemas/SsuAllocStrategy'
        tenant:
          type: string
          default: ""
          description: >-
            Requester tenant isolation identifier. Can be empty; when non-empty,
            only [a-zA-Z0-9_-.:] is allowed.
    SsuSpaceReq:
      type: object
      description: Attach/detach request for a storage space.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 48
          description: SSU storage space identifier.
          example: "ssu-volume-001"
        nqn:
          type: string
          maxLength: 69
          default: ""
          description: Host NVMe Qualified Name (NQN). Can be empty.
          example: "nqn.2014-08.org.nvmexpress:uuid:fedcba98-7654-3210-fedc-ba9876543210"
        src_eid:
          type: string
          default: ""
          maxLength: 17
          description: Source EID (Endpoint Identifier). Can be empty.
          example: "0x0000"
    SsuAttachResult:
      type: object
      description: Result of attaching a storage space.
      required:
        - dev_paths
      properties:
        dev_paths:
          type: array
          description: Device paths returned after the storage space is attached.
          items:
            type: string
            maxLength: 128
          example: ["/dev/nvme0n1"]
    SsuInfo:
      type: object
      description: SSU allocation result.
      required:
        - name
        - strategy
        - namespace_cnt
        - namespaces
      properties:
        name:
          type: string
          maxLength: 256 # Update when the actual UBS_SSU_MAX_NAME_LENGTH is available.
          description: SSU name (fixed-length string).
          example: "ssu-volume-001"
        strategy:
          type: integer
          format: int32
          description: Allocation strategy.
          example: 1
        namespace_cnt:
          type: integer
          format: int64
          minimum: 0
          description: Number of namespaces.
          example: 2
        namespaces:
          type: array
          description: Namespace information entries.
          items:
            $ref: '#/components/schemas/SsuNamespaceInfo'
    SsuNamespaceInfo:
      type: object
      description: SSU namespace details.
      required:
        - tgt_eid
        - tgt_nqn
        - ns_uuid
        - ns_id
        - ns_dev_path
        - ns_size
        - lba_format
      properties:
        tgt_eid:
          type: string
          maxLength: 64 # Update when the actual UBS_SSU_MAX_EID_LENGTH is available.
          description: Target EID (Endpoint Identifier).
          example: "nqn.2014-08.org.nvmexpress:uuid:fedcba98-7654-3210-fedc-ba9876543210"
        tgt_nqn:
          type: string
          maxLength: 256 # Update when the actual UBS_SSU_MAX_NQN_LENGTH is available.
          description: Target NQN (NVMe Qualified Name).
          example: "nqn.2014-08.com.example:nvme:subsystem-1"
        ns_uuid:
          type: string
          maxLength: 64 # Update when the actual UBS_SSU_MAX_UUID_LENGTH is available.
          description: Namespace UUID.
          example: "123e4567-e89b-12d3-a456-426614174000"
        ns_id:
          type: integer
          minimum: 0
          description: Namespace ID.
          example: 1
        ns_dev_path:
          type: string
          maxLength: 128 # Update when the actual UBS_SSU_MAX_DEV_PATH_LENGTH is available.
          description: Namespace device path in the VM.
          example: "/dev/nvme0n1"
        ns_size:
          type: integer
          minimum: 0
          description: Namespace size in bytes.
          example: 1099511627776
        lba_format:
          type: integer
          description: LBA format (Logical Block Address Format).
          example: 0
    Message:
      type: object
      description: Response message.
      required:
        - msg
      properties:
        msg:
          type: string
          maxLength: 256
          description: Returned message.
          example: "success."
...